refactor(reservation): 예약 도메인이 소유/과거 여부를 직접 판단하도록 캡슐화
